resolved bugs

+++ gps_module/gps.h
@@ -147,7 +147,7 @@ typedef struct {
     // Coordinates
     //int32_t longitude;           // Longitude (degrees * 1e7)
     double longitude;
-    int32_t latitude;            // Latitude (degrees * 1e7)
+    double latitude;            // Latitude (degrees * 1e7)
     int32_t height;              // Height above ellipsoid (millimeters)
     int32_t heightMSL;                // Height above mean sea level (millimeters)
